Start moving markup to appropriate component (in this case, textSearch); ensure all size parameters are referring to bootstrap less variables.
