+++ index.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="build/style.css">
 </head>
 <body>
-    <div id="content" class="container"></div>
+    <div id="content"></div>
     <script src="build/bundle.js"></script>
 </body>
 </html>
