+++ scripts/components/textSearch.jsx
@@ -4,8 +4,15 @@ var React = require('react');
 var QueryActions = require('../actions/queryActions');
 var _ = require('lodash');
 var filters = require('../config/filters');
+var bs = require('react-bootstrap');
+var Nav = bs.Nav,
+  Button = bs.Button,
+  Input = bs.Input;
 
 var TextSearch = React.createClass({
+  propTypes: {
+    search: React.PropTypes.object.isRequired
+  },
   handleChange: function(e) {
     var node = this.refs.searchBox.getDOMNode();
     // required for testing.
@@ -16,10 +23,29 @@ var TextSearch = React.createClass({
     QueryActions.setQueryString(queryString);
   },
   render: function(){
+    var resultsCountStatement = (
+      <small>
+        <div className="resultsCount"><strong>1634634</strong> genes</div>
+        <div><strong>38</strong> genomes</div>
+      </small>
+    );
+
+    var filterDropdown = (
+      <Button>Filter <span className="caret"></span></Button>
+    );
+
     return (
-      <div className="search">
-        <input placeholder="Search for Genes" ref="searchBox" type="text" defaultValue={this.props.query.q} onChange={this.handleChange} />
-      </div>
+      <Nav right={true} className="searchBoxNav">
+        <Input className="foo"
+               type="search"
+               placeholder="Search for genes…"
+               standalone={true}
+               addonAfter={resultsCountStatement}
+               buttonAfter={filterDropdown} />
+      </Nav>
+      //<div className="search">
+      //  <input placeholder="Search for Genes" ref="searchBox" type="text" defaultValue={this.props.query.q} onChange={this.handleChange} />
+      //</div>
     );
   }
 });
