--- styles/body.less
@@ -1,9 +1,0 @@
-@import 'colors';
-
-body {
-  font-family: 'Open Sans', sans-serif;
-  font-size: 16px;
-  margin: 0;
-  padding: 0;
-  background-color: @gramene-offwhite;
-}
