+++ styles/main.less
@@ -1,5 +1,4 @@
 @import '../node_modules/bootstrap/less/bootstrap';
 @import 'colors';
-@import 'body';
 @import 'header';
 @import 'filters';
